Refactore remove excessive console logs
