--- helpers/validateBody.js
@@ -1,15 +1,0 @@
-import HttpError from "./HttpError.js";
-
-const validateBody = (schema) => {
-  const func = (req, _, next) => {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      next(HttpError(400, error.message));
-    }
-    next();
-  };
-
-  return func;
-};
-
-export default validateBody;
